+++ projects/swimlane/ngx-datatable/src/lib/types/reorder.type.ts
@@ -1,0 +1,10 @@
+import { CdkDragDrop } from '@angular/cdk/drag-drop';
+
+/** Row */
+export interface NgxRowReorder {
+  current: any[],
+  previous: any[],
+  currentIndex: number,
+  previousIndex: number,
+  event: CdkDragDrop<any>
+}